+++ src/cow-react/modules/swap/containers/Row/RowReceivedAfterSlippage/index.tsx
