+++ src/cow-react/modules/swap/containers/TradeBasicDetails/index.tsx
@@ -4,9 +4,9 @@ import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 import { INITIAL_ALLOWED_SLIPPAGE_PERCENT } from 'constants/index'
 import { useHigherUSDValue } from 'hooks/useStablecoinPrice'
 import { LowerSectionWrapper } from '@cow/modules/swap/pure/styled'
-import { RowFee } from '@cow/modules/swap/containers/RowFee'
-import { RowSlippage } from '@cow/modules/swap/containers/RowSlippage'
-import { RowReceivedAfterSlippage } from '@cow/modules/swap/containers/RowReceivedAfterSlippage'
+import { RowFee } from '@cow/modules/swap/containers/Row/RowFee'
+import { RowSlippage } from '@cow/modules/swap/containers/Row/RowSlippage'
+import { RowReceivedAfterSlippage } from '@cow/modules/swap/containers/Row/RowReceivedAfterSlippage'
 
 interface TradeBasicDetailsProp extends BoxProps {
   allowedSlippage: Percent | string
