+++ src/cow-react/modules/swap/pure/Row/RowSlippageContent/index.cosmos.tsx
@@ -1,6 +1,6 @@
 import { Percent } from '@uniswap/sdk-core'
 import { RowSlippageContent, RowSlippageContentProps } from '@cow/modules/swap/pure/Row/RowSlippageContent'
-import { RowSlippageProps } from '@cow/modules/swap/containers/RowSlippage'
+import { RowSlippageProps } from '@cow/modules/swap/containers/Row/RowSlippage'
 
 const defaultProps: RowSlippageProps & RowSlippageContentProps = {
   isEthFlow: true,
