+++ src/cow-react/modules/swap/pure/Row/RowSlippageContent/index.tsx
@@ -5,7 +5,7 @@ import { RowFixed } from 'components/Row'
 import { MouseoverTooltipContent } from 'components/Tooltip'
 import { INPUT_OUTPUT_EXPLANATION, PERCENTAGE_PRECISION } from 'constants/index'
 import { StyledInfo } from '@cow/pages/Swap/styleds'
-import { RowSlippageProps } from '@cow/modules/swap/containers/RowSlippage'
+import { RowSlippageProps } from '@cow/modules/swap/containers/Row/RowSlippage'
 import { StyledRowBetween, TextWrapper } from '@cow/modules/swap/pure/Row/styled'
 import { RowStyleProps } from '@cow/modules/swap/pure/Row/typings'
 import { ThemedText } from 'theme/index'
