+++ src/cow-react/modules/swap/pure/TradeRates/index.tsx
@@ -7,7 +7,7 @@ import { ApplicationModal } from 'state/application/reducer'
 import { Price } from '@cow/modules/swap/pure/Price'
 import TradeGp from 'state/swap/TradeGp'
 import { INITIAL_ALLOWED_SLIPPAGE_PERCENT } from 'constants/index'
-import { RowSlippage } from '@cow/modules/swap/containers/RowSlippage'
+import { RowSlippage } from '@cow/modules/swap/containers/Row/RowSlippage'
 import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 import { genericPropsChecker } from '@cow/modules/swap/containers/NewSwapWidget/propsChecker'
 import { TradeBasicDetails } from '@cow/modules/swap/containers/TradeBasicDetails'
